+++ lib/auth.rb
