+++ lib/main.rb
@@ -1,14 +1,14 @@
-require 'http_user'
-require 'http_auth'
-require 'http_task'
-require 'http_task_daily'
+require 'user_api'
+require 'auth_api'
+require 'task_api'
+require 'task_daily_api'
 
 module Main
   # main api
   class Api < Grape::API
     format :json
 
-    mount Auth::Auth
+    mount Auth::Api
 
     namespace 'api' do
       mount Api::User
